# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "Sinapsistencia API", "description": "Plataforma de mediación médico-legal — backend Spring Boot. Contrato de respuesta: {success, data} / {success, error}.", "version": "v1"}, "servers": [{"url": "http://localhost:8080", "description": "Generated server url"}], "paths": {"/api/users/{id}": {"get": {"tags": ["user-controller"], "operationId": "get", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseUserDto"}}}}}}, "put": {"tags": ["user-controller"], "operationId": "update", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateUserRequest"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseUserDto"}}}}}}, "patch": {"tags": ["user-controller"], "operationId": "toggleActive", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseUserDto"}}}}}}}, "/api/profile": {"get": {"tags": ["profile-controller"], "operationId": "get_1", "parameters": [{"name": "userId", "in": "query", "required": false, "schema": {"type": "string"}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseProfileResponse"}}}}}}, "put": {"tags": ["profile-controller"], "operationId": "put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateProfileRequest"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseMessageResponse"}}}}}}, "patch": {"tags": ["profile-controller"], "operationId": "patch", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateProfileRequest"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseMessageResponse"}}}}}}}, "/api/legal-cases/{id}": {"get": {"tags": ["legal-case-controller"], "operationId": "get_2", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseCaseResponse"}}}}}}, "put": {"tags": ["legal-case-controller"], "operationId": "update_1", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateCaseRequest"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseCaseResponse"}}}}}}}, "/api/documents/{id}": {"get": {"tags": ["document-controller"], "operationId": "get_3", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseDocumentResponse"}}}}}}, "put": {"tags": ["document-controller"], "operationId": "update_2", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateDocumentRequest"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseDocumentResponse"}}}}}}}, "/api/users": {"get": {"tags": ["user-controller"], "operationId": "list", "parameters": [{"name": "role", "in": "query", "required": false, "schema": {"type": "string"}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseListUserDto"}}}}}}, "post": {"tags": ["user-controller"], "operationId": "create", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateUserRequest"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseUserDto"}}}}}}}, "/api/profile/avatar": {"post": {"tags": ["profile-controller"], "operationId": "uploadAvatar", "parameters": [{"name": "userId", "in": "query", "required": false, "schema": {"type": "string"}}], "requestBody": {"content": {"application/json": {"schema": {"type": "object", "properties": {"file": {"type": "string", "format": "binary"}}, "required": ["file"]}}}}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseMapStringString"}}}}}}}, "/api/ml/risk": {"post": {"tags": ["ml-controller"], "operationId": "risk", "requestBody": {"content": {"application/json": {"schema": {"type": "object", "additionalProperties": {"type": "object"}}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseMapStringObject"}}}}}}}, "/api/matching/lawyers": {"get": {"tags": ["matching-controller"], "operationId": "lawyers", "parameters": [{"name": "doctorId", "in": "query", "required": false, "schema": {"type": "string"}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseObject"}}}}}}, "post": {"tags": ["matching-controller"], "operationId": "generateRecommendations", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/GenerateRecommendationsBody"}}}}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseRecommendationsResponse"}}}}}}}, "/api/matching/contact-requests": {"get": {"tags": ["matching-controller"], "operationId": "contactRequests", "parameters": [{"name": "lawyerId", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "doctorId", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "status", "in": "query", "required": false, "schema": {"type": "string"}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseListContactRequestResponse"}}}}}}, "post": {"tags": ["matching-controller"], "operationId": "createContactRequest", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateContactRequestBody"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseContactRequestResponse"}}}}}}, "patch": {"tags": ["matching-controller"], "operationId": "respondContactRequest", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RespondContactRequestBody"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseContactRequestResponse"}}}}}}}, "/api/legal-cases": {"get": {"tags": ["legal-case-controller"], "operationId": "list_1", "parameters": [{"name": "status", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "priority", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "doctorId", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "search", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "format": "int32", "default": 1}}, {"name": "pageSize", "in": "query", "required": false, "schema": {"type": "integer", "format": "int32", "default": 20}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseListResponseCaseResponse"}}}}}}, "post": {"tags": ["legal-case-controller"], "operationId": "create_1", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateCaseRequest"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseCaseResponse"}}}}}}}, "/api/documents": {"get": {"tags": ["document-controller"], "operationId": "list_2", "parameters": [{"name": "status", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "type", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "authorId", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "search", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "format": "int32", "default": 1}}, {"name": "pageSize", "in": "query", "required": false, "schema": {"type": "integer", "format": "int32", "default": 20}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseListResponseDocumentResponse"}}}}}}, "post": {"tags": ["document-controller"], "operationId": "create_2", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateDocumentRequest"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseDocumentResponse"}}}}}}}, "/api/auth/register": {"post": {"tags": ["auth-controller"], "operationId": "register", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/RegisterRequest"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseMessageResponse"}}}}}}}, "/api/auth/logout": {"post": {"tags": ["auth-controller"], "operationId": "logout", "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseMessageResponse"}}}}}}}, "/api/auth/login": {"post": {"tags": ["auth-controller"], "operationId": "login", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseLoginResponse"}}}}}}}, "/api/ml/health": {"get": {"tags": ["ml-controller"], "operationId": "health", "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseMapStringObject"}}}}}}}, "/api/matching/relevant-cases": {"get": {"tags": ["matching-controller"], "operationId": "relevantCases", "parameters": [{"name": "lawyerId", "in": "query", "required": false, "schema": {"type": "string"}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseMapStringObject"}}}}}}}, "/api/matching/doctors": {"get": {"tags": ["matching-controller"], "operationId": "doctors", "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseListDoctorCardDto"}}}}}}}, "/api/auth/me": {"get": {"tags": ["auth-controller"], "operationId": "me", "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseUserDto"}}}}}}}, "/api/audit": {"get": {"tags": ["audit-controller"], "operationId": "list_3", "parameters": [{"name": "action", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "userId", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "resource", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "search", "in": "query", "required": false, "schema": {"type": "string"}}, {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "format": "int32", "default": 1}}, {"name": "pageSize", "in": "query", "required": false, "schema": {"type": "integer", "format": "int32", "default": 50}}], "responses": {"200": {"description": "OK", "content": {"*/*": {"schema": {"$ref": "#/components/schemas/ApiResponseListResponseAuditLogResponse"}}}}}}}}, "components": {"schemas": {"UpdateUserRequest": {"type": "object", "properties": {"name": {"type": "string"}, "role": {"type": "string"}}}, "ApiResponseUserDto": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"$ref": "#/components/schemas/UserDto"}, "error": {"type": "string"}}}, "UserDto": {"type": "object", "properties": {"id": {"type": "string"}, "email": {"type": "string"}, "name": {"type": "string"}, "role": {"type": "string"}, "isActive": {"type": "boolean"}, "createdAt": {"type": "string", "format": "date-time"}, "avatar": {"type": "string"}}}, "ProfessionalPayload": {"type": "object", "properties": {"cmp": {"type": "string"}, "specialty": {"type": "string"}, "hospital": {"type": "string"}, "cab": {"type": "string"}, "specialties": {"type": "array", "items": {"type": "string"}}, "medicalAreas": {"type": "array", "items": {"type": "string"}}, "yearsExperience": {"type": "integer", "format": "int32"}, "available": {"type": "boolean"}, "phone": {"type": "string"}, "bio": {"type": "string"}}}, "UpdateProfileRequest": {"type": "object", "properties": {"userId": {"type": "string"}, "name": {"type": "string"}, "professional": {"$ref": "#/components/schemas/ProfessionalPayload"}}}, "ApiResponseMessageResponse": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"$ref": "#/components/schemas/MessageResponse"}, "error": {"type": "string"}}}, "MessageResponse": {"type": "object", "properties": {"message": {"type": "string"}}}, "UpdateCaseRequest": {"type": "object", "properties": {"status": {"type": "string"}, "lawyerId": {"type": "string"}, "notes": {"type": "string"}}}, "ApiResponseCaseResponse": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"$ref": "#/components/schemas/CaseResponse"}, "error": {"type": "string"}}}, "CaseContextDto": {"type": "object", "properties": {"id": {"type": "string"}, "contextCode": {"type": "string"}, "ageReference": {"type": "integer", "format": "int32"}, "medicalArea": {"type": "string"}, "eventDate": {"type": "string", "format": "date"}, "summary": {"type": "string"}, "relevantFactors": {"type": "array", "items": {"type": "string"}}, "isSimulated": {"type": "boolean"}}}, "CaseResponse": {"type": "object", "properties": {"id": {"type": "string"}, "title": {"type": "string"}, "description": {"type": "string"}, "status": {"type": "string"}, "priority": {"type": "string"}, "priorityJustification": {"type": "string"}, "medicalSpecialty": {"type": "string"}, "eventType": {"type": "string"}, "perceivedUrgency": {"type": "string"}, "doctorId": {"type": "string"}, "doctor": {"$ref": "#/components/schemas/PersonSnapshot"}, "lawyerId": {"type": "string"}, "lawyer": {"$ref": "#/components/schemas/PersonSnapshot"}, "context": {"$ref": "#/components/schemas/CaseContextDto"}, "notes": {"type": "string"}, "createdAt": {"type": "string", "format": "date-time"}, "updatedAt": {"type": "string", "format": "date-time"}, "procedureComplexity": {"type": "string"}, "complexitySource": {"type": "string"}, "documentationComplete": {"type": "boolean"}, "informedConsent": {"type": "boolean"}, "hasPriorComplaints": {"type": "boolean"}}}, "PersonSnapshot": {"type": "object", "properties": {"id": {"type": "string"}, "fullName": {"type": "string"}, "email": {"type": "string"}}}, "UpdateDocumentRequest": {"type": "object", "properties": {"status": {"type": "string"}}}, "ApiResponseDocumentResponse": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"$ref": "#/components/schemas/DocumentResponse"}, "error": {"type": "string"}}}, "DocumentResponse": {"type": "object", "properties": {"id": {"type": "string"}, "title": {"type": "string"}, "type": {"type": "string"}, "status": {"type": "string"}, "caseId": {"type": "string"}, "authorId": {"type": "string"}, "authorName": {"type": "string"}, "currentVersionId": {"type": "string"}, "versions": {"type": "array", "items": {"$ref": "#/components/schemas/VersionDto"}}, "signatures": {"type": "array", "items": {"$ref": "#/components/schemas/SignatureDto"}}, "createdAt": {"type": "string", "format": "date-time"}, "updatedAt": {"type": "string", "format": "date-time"}}}, "SignatureDto": {"type": "object", "properties": {"id": {"type": "string"}, "signerId": {"type": "string"}, "type": {"type": "string"}, "signedAt": {"type": "string", "format": "date-time"}, "isValid": {"type": "boolean"}, "hash": {"type": "string"}}}, "VersionDto": {"type": "object", "properties": {"id": {"type": "string"}, "version": {"type": "integer", "format": "int32"}, "content": {"type": "string"}, "fileUrl": {"type": "string"}, "createdById": {"type": "string"}, "createdAt": {"type": "string", "format": "date-time"}, "notes": {"type": "string"}}}, "CreateUserRequest": {"type": "object", "properties": {"name": {"type": "string"}, "email": {"type": "string"}, "role": {"type": "string"}, "password": {"type": "string"}}}, "ApiResponseMapStringString": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"type": "object", "additionalProperties": {"type": "string"}}, "error": {"type": "string"}}}, "ApiResponseMapStringObject": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"type": "object", "additionalProperties": {"type": "object"}}, "error": {"type": "string"}}}, "GenerateRecommendationsBody": {"type": "object", "properties": {"caseId": {"type": "string"}}}, "ApiResponseRecommendationsResponse": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"$ref": "#/components/schemas/RecommendationsResponse"}, "error": {"type": "string"}}}, "JsonNode": {"type": "object"}, "LawyerCardDto": {"type": "object", "properties": {"id": {"type": "string"}, "userId": {"type": "string"}, "fullName": {"type": "string"}, "email": {"type": "string"}, "cab": {"type": "string"}, "specialties": {"type": "array", "items": {"type": "string"}}, "medicalAreas": {"type": "array", "items": {"type": "string"}}, "yearsExperience": {"type": "integer", "format": "int32"}, "resolvedCases": {"type": "integer", "format": "int32"}, "rating": {"type": "number", "format": "double"}, "phone": {"type": "string"}, "bio": {"type": "string"}, "available": {"type": "boolean"}, "avatar": {"type": "string"}}}, "RecommendationDto": {"type": "object", "properties": {"id": {"type": "string"}, "doctorId": {"type": "string"}, "lawyer": {"$ref": "#/components/schemas/LawyerCardDto"}, "score": {"type": "integer", "format": "int32"}, "contentScore": {"type": "integer", "format": "int32"}, "performanceScore": {"type": "integer", "format": "int32"}, "collaborativeScore": {"type": "integer", "format": "int32"}, "scoreRaw": {"type": "number", "format": "double"}, "contentScoreRaw": {"type": "number", "format": "double"}, "performanceScoreRaw": {"type": "number", "format": "double"}, "matchedSpecialties": {"type": "array", "items": {"type": "string"}}, "modelUsed": {"type": "string"}, "featureImportance": {"$ref": "#/components/schemas/JsonNode"}, "reasons": {"type": "array", "items": {"type": "string"}}, "createdAt": {"type": "string"}}}, "RecommendationsResponse": {"type": "object", "properties": {"recommendations": {"type": "array", "items": {"$ref": "#/components/schemas/RecommendationDto"}}, "modelInfo": {"type": "object", "additionalProperties": {"type": "object"}}, "advisoryNote": {"type": "string"}}}, "CreateContactRequestBody": {"type": "object", "properties": {"fromDoctorId": {"type": "string"}, "toLawyerId": {"type": "string"}, "message": {"type": "string"}, "caseId": {"type": "string"}}}, "ApiResponseContactRequestResponse": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"$ref": "#/components/schemas/ContactRequestResponse"}, "error": {"type": "string"}}}, "ContactRequestResponse": {"type": "object", "properties": {"id": {"type": "string"}, "fromDoctorId": {"type": "string"}, "fromDoctor": {"$ref": "#/components/schemas/DoctorSnapshot"}, "toLawyerId": {"type": "string"}, "toLawyer": {"$ref": "#/components/schemas/LawyerSnapshot"}, "caseId": {"type": "string"}, "caseTitle": {"type": "string"}, "status": {"type": "string"}, "message": {"type": "string"}, "responseMessage": {"type": "string"}, "mlScore": {"type": "number", "format": "double"}, "createdAt": {"type": "string", "format": "date-time"}, "respondedAt": {"type": "string", "format": "date-time"}}}, "DoctorSnapshot": {"type": "object", "properties": {"id": {"type": "string"}, "userId": {"type": "string"}, "fullName": {"type": "string"}, "email": {"type": "string"}, "cmp": {"type": "string"}, "specialty": {"type": "string"}, "hospital": {"type": "string"}, "phone": {"type": "string"}}}, "LawyerSnapshot": {"type": "object", "properties": {"id": {"type": "string"}, "userId": {"type": "string"}, "fullName": {"type": "string"}, "email": {"type": "string"}, "cab": {"type": "string"}, "specialties": {"type": "array", "items": {"type": "string"}}, "phone": {"type": "string"}, "rating": {"type": "number", "format": "double"}}}, "ContextPayload": {"type": "object", "properties": {"contextCode": {"type": "string"}, "ageReference": {"type": "integer", "format": "int32"}, "medicalArea": {"type": "string"}, "eventDate": {"type": "string", "format": "date"}, "summary": {"type": "string"}, "relevantFactors": {"type": "array", "items": {"type": "string"}}}}, "CreateCaseRequest": {"type": "object", "properties": {"title": {"type": "string"}, "description": {"type": "string"}, "priority": {"type": "string"}, "medicalSpecialty": {"type": "string"}, "eventType": {"type": "string"}, "perceivedUrgency": {"type": "string"}, "notes": {"type": "string"}, "context": {"$ref": "#/components/schemas/ContextPayload"}, "procedureComplexity": {"type": "string"}}}, "CreateDocumentRequest": {"type": "object", "properties": {"title": {"type": "string"}, "type": {"type": "string"}, "caseId": {"type": "string"}, "initialContent": {"type": "string"}}}, "RegisterRequest": {"type": "object", "properties": {"name": {"type": "string"}, "email": {"type": "string"}, "password": {"type": "string"}, "role": {"type": "string"}, "cmp": {"type": "string"}, "specialty": {"type": "string"}, "hospital": {"type": "string"}, "cab": {"type": "string"}, "legalSpecialties": {"type": "array", "items": {"type": "string"}}, "medicalAreas": {"type": "array", "items": {"type": "string"}}}}, "LoginRequest": {"type": "object", "properties": {"email": {"type": "string"}, "password": {"type": "string"}, "role": {"type": "string"}}}, "ApiResponseLoginResponse": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"$ref": "#/components/schemas/LoginResponse"}, "error": {"type": "string"}}}, "LoginResponse": {"type": "object", "properties": {"user": {"$ref": "#/components/schemas/UserDto"}, "token": {"type": "string"}}}, "RespondContactRequestBody": {"type": "object", "properties": {"requestId": {"type": "string"}, "status": {"type": "string"}, "responseMessage": {"type": "string"}}}, "ApiResponseListUserDto": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"type": "array", "items": {"$ref": "#/components/schemas/UserDto"}}, "error": {"type": "string"}}}, "ApiResponseProfileResponse": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"$ref": "#/components/schemas/ProfileResponse"}, "error": {"type": "string"}}}, "ProfessionalDto": {"type": "object", "properties": {"cmp": {"type": "string"}, "specialty": {"type": "string"}, "subSpecialties": {"type": "array", "items": {"type": "string"}}, "hospital": {"type": "string"}, "languages": {"type": "array", "items": {"type": "string"}}, "cab": {"type": "string"}, "specialties": {"type": "array", "items": {"type": "string"}}, "medicalAreas": {"type": "array", "items": {"type": "string"}}, "rating": {"type": "number"}, "resolvedCases": {"type": "integer", "format": "int32"}, "available": {"type": "boolean"}, "yearsExperience": {"type": "integer", "format": "int32"}, "phone": {"type": "string"}, "bio": {"type": "string"}}}, "ProfileResponse": {"type": "object", "properties": {"id": {"type": "string"}, "email": {"type": "string"}, "name": {"type": "string"}, "role": {"type": "string"}, "avatar": {"type": "string"}, "isActive": {"type": "boolean"}, "createdAt": {"type": "string", "format": "date-time"}, "professional": {"$ref": "#/components/schemas/ProfessionalDto"}}}, "ApiResponseObject": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"type": "object"}, "error": {"type": "string"}}}, "ApiResponseListDoctorCardDto": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"type": "array", "items": {"$ref": "#/components/schemas/DoctorCardDto"}}, "error": {"type": "string"}}}, "DoctorCardDto": {"type": "object", "properties": {"id": {"type": "string"}, "userId": {"type": "string"}, "fullName": {"type": "string"}, "email": {"type": "string"}, "cmp": {"type": "string"}, "specialty": {"type": "string"}, "subSpecialties": {"type": "array", "items": {"type": "string"}}, "hospital": {"type": "string"}, "yearsExperience": {"type": "integer", "format": "int32"}, "phone": {"type": "string"}, "bio": {"type": "string"}, "languages": {"type": "array", "items": {"type": "string"}}, "available": {"type": "boolean"}, "avatar": {"type": "string"}}}, "ApiResponseListContactRequestResponse": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"type": "array", "items": {"$ref": "#/components/schemas/ContactRequestResponse"}}, "error": {"type": "string"}}}, "ApiResponseListResponseCaseResponse": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"$ref": "#/components/schemas/ListResponseCaseResponse"}, "error": {"type": "string"}}}, "ListResponseCaseResponse": {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/CaseResponse"}}, "total": {"type": "integer", "format": "int64"}, "page": {"type": "integer", "format": "int32"}, "pageSize": {"type": "integer", "format": "int32"}, "totalPages": {"type": "integer", "format": "int32"}}}, "ApiResponseListResponseDocumentResponse": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"$ref": "#/components/schemas/ListResponseDocumentResponse"}, "error": {"type": "string"}}}, "ListResponseDocumentResponse": {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/DocumentResponse"}}, "total": {"type": "integer", "format": "int64"}, "page": {"type": "integer", "format": "int32"}, "pageSize": {"type": "integer", "format": "int32"}, "totalPages": {"type": "integer", "format": "int32"}}}, "ApiResponseListResponseAuditLogResponse": {"type": "object", "properties": {"success": {"type": "boolean"}, "data": {"$ref": "#/components/schemas/ListResponseAuditLogResponse"}, "error": {"type": "string"}}}, "AuditLogResponse": {"type": "object", "properties": {"id": {"type": "string"}, "userId": {"type": "string"}, "userName": {"type": "string"}, "userRole": {"type": "string"}, "action": {"type": "string"}, "resource": {"type": "string"}, "resourceId": {"type": "string"}, "description": {"type": "string"}, "details": {"$ref": "#/components/schemas/JsonNode"}, "ipAddress": {"type": "string"}, "createdAt": {"type": "string", "format": "date-time"}}}, "ListResponseAuditLogResponse": {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/AuditLogResponse"}}, "total": {"type": "integer", "format": "int64"}, "page": {"type": "integer", "format": "int32"}, "pageSize": {"type": "integer", "format": "int32"}, "totalPages": {"type": "integer", "format": "int32"}}}}}}